Separate logic from base GrainReference
In preparation to move GrainReference to an Abstractions assembly, but still keep the logic inside the Orleans implementation.
- GrainReference is now bound to IGrainReferenceRuntime instead of IRuntimeClient. The new runtime abstraction includes all the logic that was previously part of GrainReference, and it's a much more reduced surface area compared to IRuntimeClient, to expose to the public.
- Moved Orleans-specific serialization outside the base abstraction.
