+++ src/OrleansRuntime/Catalog/ActivationData.cs
@@ -204,7 +204,7 @@ namespace Orleans.Runtime
             }
             CollectionAgeLimit = ageLimit;
 
-            GrainReference = GrainReference.FromGrainId(addr.Grain, runtimeClient, genericArguments, Grain.IsSystemTarget ? addr.Silo : null);
+            GrainReference = GrainReference.FromGrainId(addr.Grain, runtimeClient.GrainReferenceRuntime, genericArguments, Grain.IsSystemTarget ? addr.Silo : null);
             this.SchedulingContext = new SchedulingContext(this);
         }
